requirements.txt
Django==5.1.3
django-ckeditor
django-cleanup
django-crispy-forms